+++ changelog.md
@@ -2,4 +2,5 @@
 
 ### 0.1.x
 - 0.1.0
-  - Added `Calcite Wall`
+  - Added new `Walls` variants
+    - `Calcite Wall`