+++ changelog.md
@@ -4,3 +4,7 @@
 - 0.1.0
   - Added new `Walls` variants
     - `Calcite Wall`
+    - `Polished Basalt`
+    - `Smooth Basalt`
+  - Added new `Stairs` variants
+    - `Calcite Stairs`